refactor dynamic-separator. create save.js.
